# backend/internal/platform/httpx/api/openapi.gen.yaml
openapi: 3.1.0
info:
  title: PAO API
  version: 1.0.0
  license:
    name: Proprietary — all rights reserved
    url: https://getpao.com/legal
  description: |
    REST contract shared by the PAO Go backend and the Flutter apps.

    - Versioned under `/v1`, with route groups `/v1/customer`, `/v1/provider` and `/v1/admin` (PRD §9.6).
    - Every error uses `ErrorResponse`; `code` is a stable `ErrorCode` the apps translate.
    - `x-permission` on each operation names the permission the caller must hold
      (`public` = no token, `authenticated` = any signed-in account). `:own` permissions
      are also checked against the resource owner.
    - Lists use cursor pagination (`cursor`, `limit` ≤ 100, `nextCursor`).
    - Money is integer paisa. Times are UTC (RFC 3339); apps display Asia/Dhaka.
servers:
  - url: http://localhost:8080
    description: Local Docker stack
security:
  - bearerAuth: []
...
paths:
  /v1/auth/otp/request:
    post:
      operationId: requestOtp
      tags:
        - Auth
      summary: Send a 6-digit sign-in code by SMS
      x-permission: public
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpRequestAccepted'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/auth/otp/verify:
    post:
      operationId: verifyOtp
      tags:
        - Auth
      summary: Verify the code and sign in (creates the account on first use)
      x-permission: public
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerifyRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags:
        - Auth
      summary: Rotate the refresh token and issue a new access token
      x-permission: public
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
          headers:
            Set-Cookie:
              description: 'Admin web only: the rotated refresh token as an HttpOnly, Secure, SameSite=Strict cookie.'
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
      parameters:
        - name: pao_refresh
          in: cookie
          required: false
          description: Admin web refresh token (HttpOnly cookie).
          schema:
            type: string
...
  /v1/auth/admin/login:
    post:
      operationId: adminLogin
      tags:
        - Auth
      summary: 'Admin step 1: email and password; returns a TOTP challenge'
      x-permission: public
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginChallenge'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/auth/admin/totp:
    post:
      operationId: adminVerifyTotp
      tags:
        - Auth
      summary: 'Admin step 2: TOTP code; sets the refresh cookie'
      x-permission: public
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminTotpRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
          headers:
            Set-Cookie:
              description: 'Admin web only: the rotated refresh token as an HttpOnly, Secure, SameSite=Strict cookie.'
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/auth/admin/password:
    post:
      operationId: changeAdminPassword
      tags:
        - Auth
      summary: 'Admin: replace the temporary or current password'
      x-permission: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPasswordChange'
      responses:
        '204':
          description: No content
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: EdDSA access token (15 minutes).
    refreshCookie:
      type: apiKey
      in: cookie
      name: pao_refresh
      description: Admin web refresh token (HttpOnly, Secure, SameSite=Strict).
...
  responses:
    Unauthenticated:
      description: Missing, invalid or expired access token (UNAUTHENTICATED, TOKEN_EXPIRED) or a failed credential check.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHENTICATED
              message: Sign in again.
    Forbidden:
      description: The caller lacks the operation's x-permission, does not own the resource, or the account is banned or suspended.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: Permission booking:cancel:own required.
...
    ValidationFailed:
      description: The request does not match the schema or a business validation rule.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_FAILED
              message: Invalid request.
              details:
                field: phone
    TooManyRequests:
      description: Rate limit or attempt limit reached; retry after the given number of seconds.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: RATE_LIMITED
              message: Too many requests.
  schemas:
    ErrorResponse:
      type: object
      description: The single error format used by every endpoint (PRD §9.6).
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      example:
        error:
          code: NOT_FOUND
          message: Booking not found.
    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: English description for logs; apps translate `code` instead.
        details:
          type: object
          description: Machine-readable context, e.g. the invalid fields.
          additionalProperties: true
    ErrorCode:
      type: string
      description: Stable error codes the apps translate (04-decisions.md E11).
      enum:
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - TOKEN_EXPIRED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - INTERNAL
        - NOT_IMPLEMENTED
        - OTP_INVALID
        - OTP_EXPIRED
        - OTP_LOCKED
        - ACCOUNT_BANNED
        - ACCOUNT_SUSPENDED
        - INVALID_CREDENTIALS
        - ACCOUNT_LOCKED
        - TOTP_INVALID
        - MFA_CHALLENGE_EXPIRED
        - REFRESH_TOKEN_INVALID
        - IDEMPOTENCY_KEY_REQUIRED
        - IDEMPOTENCY_KEY_REUSED
        - ADDRESS_LIMIT_REACHED
        - OUTSIDE_SERVICE_AREA
        - SERVICE_UNAVAILABLE
        - PROVIDER_UNAVAILABLE
        - DUPLICATE_BOOKING_REQUEST
        - BOOKING_INVALID_TRANSITION
        - BOOKING_ALREADY_RESPONDED
        - ACTIVE_JOB_EXISTS
        - START_CODE_INVALID
        - START_CODE_LOCKED
        - CANCELLATION_NOT_ALLOWED
        - EXTRA_ITEM_INVALID
        - EXTRAS_PENDING
        - CASH_CONFIRMATION_REQUIRED
        - REVIEW_NOT_ALLOWED
        - REVIEW_ALREADY_SUBMITTED
        - UPLOAD_INVALID
        - UPLOAD_NOT_FOUND
        - ENROLMENT_INCOMPLETE
        - NOT_VERIFIED
        - AGE_REQUIREMENT
        - LEVEL2_COOLING_OFF
        - COMPLAINT_INVALID_TRANSITION
        - SETTING_INVALID
        - ROLE_INVALID
...
    Phone:
      type: string
      description: Bangladeshi mobile number; the API normalises to +8801XXXXXXXXX.
      pattern: ^(\+?880|0)?1[3-9][0-9]{8}$
      example: '+8801712345678'
...
    OtpPurpose:
      type: string
      description: Why the code is sent. `delete_account` re-confirms account deletion.
      enum:
        - login
        - delete_account
      example: login
    OtpRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          $ref: '#/components/schemas/Phone'
        purpose:
          $ref: '#/components/schemas/OtpPurpose'
      example:
        phone: '+8801712345678'
        purpose: login
    OtpRequestAccepted:
      type: object
      required:
        - expiresInSeconds
        - resendAfterSeconds
      properties:
        expiresInSeconds:
          type: integer
          example: 300
        resendAfterSeconds:
          type: integer
          example: 60
    AppKind:
      type: string
      description: The app signing in; signing in to the partner app adds the provider role (PRD §3).
      enum:
        - customer
        - partner
      example: customer
    OtpVerifyRequest:
      type: object
      required:
        - phone
        - code
        - app
      properties:
        phone:
          $ref: '#/components/schemas/Phone'
        code:
          type: string
          pattern: ^[0-9]{6}$
          example: '482913'
        app:
          $ref: '#/components/schemas/AppKind'
    TokenPair:
      type: object
      required:
        - accessToken
        - expiresInSeconds
        - account
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJFZERTQSIsImtpZCI6ImsxIn0...
        refreshToken:
          type: string
          description: Omitted for the admin web app, which receives it as an HttpOnly cookie.
          example: 3q2-7wRk5sQ0x9nVb...
        expiresInSeconds:
          type: integer
          example: 900
        isNewAccount:
          type: boolean
          example: false
        account:
          $ref: '#/components/schemas/Account'
    RefreshRequest:
      type: object
      description: Mobile apps send the token in the body; the admin web app sends none and relies on the cookie.
      properties:
        refreshToken:
          type: string
          example: 3q2-7wRk5sQ0x9nVb...
...
    AdminLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          example: verifier@pao.bd
        password:
          type: string
          minLength: 8
          maxLength: 128
          example: correct-horse-battery
    AdminLoginChallenge:
      type: object
      required:
        - challengeId
        - expiresInSeconds
      properties:
        challengeId:
          type: string
          format: uuid
          example: 0192f4a6-1c2b-7c1d-9e3f-5a6b7c8d9e0f
        expiresInSeconds:
          type: integer
          example: 300
        totpEnrolment:
          $ref: '#/components/schemas/TotpEnrolment'
        mustChangePassword:
          type: boolean
          description: True while the admin still uses the temporary password from the invite.
    TotpEnrolment:
      type: object
      description: Present on the first login; the admin scans it before entering a code.
      required:
        - secret
        - otpauthUrl
      properties:
        secret:
          type: string
          example: JBSWY3DPEHPK3PXP
        otpauthUrl:
          type: string
          example: otpauth://totp/PAO:verifier@pao.bd?secret=JBSWY3DPEHPK3PXP&issuer=PAO
    AdminTotpRequest:
      type: object
      required:
        - challengeId
        - code
      properties:
        challengeId:
          type: string
          format: uuid
          example: 0192f4a6-1c2b-7c1d-9e3f-5a6b7c8d9e0f
        code:
          type: string
          pattern: ^[0-9]{6}$
          example: '123456'
    AdminPasswordChange:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          minLength: 8
          maxLength: 128
        newPassword:
          type: string
          minLength: 12
          maxLength: 128
          example: a-much-longer-passphrase
    AccountStatus:
      type: string
      description: PRD §6.4 account lifecycle.
      enum:
        - pending
        - active
        - suspended
        - banned
      example: active
    Role:
      type: string
      enum:
        - customer
        - provider
        - verifier
        - catalog_manager
        - support_agent
        - super_admin
      example: customer
    Account:
      type: object
      required:
        - id
        - roles
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: 0192f4a6-1c2b-7c1d-9e3f-5a6b7c8d9e0f
        phone:
          type: string
          example: '+8801712345678'
        email:
          type: string
          format: email
          example: verifier@pao.bd
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        status:
          $ref: '#/components/schemas/AccountStatus'
        createdAt:
          type: string
          format: date-time
...